+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Bug in vis.put_text with wrong padding handling
 ## Added
 - Possibility to overwrite YAML config settings from the command line
+- Face detection example with Dlib frontal face detector
 
 # 0.1.1
 Initial release